Fix AltStackInstances check for creation
